pyproject.toml
[tool.poetry]
name = "e2e_automation_test"
version = "0.1.0"
description = ""
authors = ["lebathien"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.16.0"
behavex = "3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
